fix: move energy_from_rho_P into header file: avoid relocatable dev code
