+++ src/pgen/pgen.hpp
@@ -22,6 +22,7 @@ using namespace parthenon::package::prelude;
 // internal includes
 #include "fluid/fluid.hpp"
 #include "geometry/geometry.hpp"
+#include "phoebus_utils/root_find.hpp"
 #include "phoebus_utils/variables.hpp"
 #include "radiation/radiation.hpp"
 #include "tracers/tracers.hpp"
@@ -117,9 +118,34 @@ static std::map<std::string, std::function<void(ParameterInput *pin, Mesh *pmesh
 // END OF UNTOUCHABLE MACRO SECTION
 */
 
-KOKKOS_FUNCTION
-Real energy_from_rho_P(const Microphysics::EOS::EOS &eos, const Real rho, const Real P,
-                       const Real emin, const Real emax, const Real Ye = 0.0);
+class PressResidual {
+ public:
+  KOKKOS_INLINE_FUNCTION
+  PressResidual(const EOS &eos, const Real rho, const Real P, const Real Ye)
+      : eos_(eos), rho_(rho), P_(P) {
+    lambda_[0] = Ye;
+  }
+  KOKKOS_INLINE_FUNCTION
+  Real operator()(const Real e) {
+    return eos_.PressureFromDensityInternalEnergy(rho_, e, lambda_) - P_;
+  }
+
+ private:
+  const EOS &eos_;
+  Real rho_, P_;
+  Real lambda_[2];
+};
+
+template <typename T>
+KOKKOS_FUNCTION Real energy_from_rho_P(T &eos, const Real rho, const Real P,
+                                       const Real emin, const Real emax,
+                                       const Real Ye = 0.0) {
+  PARTHENON_REQUIRE(P >= 0, "Pressure is negative!");
+  PressResidual res(eos, rho, P, Ye);
+  root_find::RootFind root;
+  Real eroot = root.regula_falsi(res, emin, emax, 1.e-10 * P, emin - 1.e10);
+  return rho * eroot;
+}
 
 } // namespace phoebus
